% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiralplot.R
\name{spiral_ozone}
\alias{spiral_ozone}
\title{Spiral plot of daily total column ozone}
\usage{
spiral_ozone(data, station, years, ring_spacing = 2)
}
\arguments{
\item{data}{A data frame containing WOUDC ozone data. The data frame must
contain the variables \code{station_name}, \code{daily_date}, and
\code{daily_columno3}.}

\item{station}{A character string specifying the station to plot.}

\item{years}{A numeric vector specifying the years to include in the plot.}

\item{ring_spacing}{A numeric value specifying the radial spacing between
successive yearly rings. Defaults to \code{2}.}
}
\value{
A ggplot object.
}
\description{
Creates a spiral time-series visualization of daily total column ozone
observations for a selected station and period of years. Each complete
circle represents one year, with days of the year mapped around the circle
and ozone values represented by colour. The radial spacing between years
can be controlled using \code{ring_spacing}.
}
\examples{
\dontrun{
spiral_ozone(
  ozonedata,
  station = "Irene",
  years = 2015:2016,
  ring_spacing = 2
)
}

}
